+++ web/src/services/theme/index.ts
@@ -7,7 +7,7 @@ export { default } from "./darkTheme";
 export { default as Provider } from "./provider";
 export { default as PublishedAppProvider } from "./publishedAppProvider";
 export { usePublishTheme, publishTheme, mask, type SceneThemeOptions } from "./publishTheme";
-export { default as GlobalStyles } from "./common/globalstyle";
+export { default as GlobalStyles } from "./common/globalStyles";
 export { default as lightTheme } from "./lightTheme";
 export { default as darkTheme } from "./darkTheme";
