+++ web/src/services/theme/publishedAppProvider.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from "@emotion/react";
 import { ReactNode } from "react";
 
-import GlobalStyle from "./common/globalstyle";
+import GlobalStyle from "./common/globalStyles";
 import theme from "./darkTheme";
 
 const Provider: React.FC<{ children?: ReactNode }> = ({ children }) => {
